Fixed support for docker compose by allowing connect/disconnect on stopped containers
