+++ daemon/container_operations_windows.go
@@ -2,25 +2,10 @@
 
 package daemon
 
-import (
-	"fmt"
-
-	networktypes "github.com/docker/docker/api/types/network"
-	"github.com/docker/docker/container"
-)
+import "github.com/docker/docker/container"
 
 func (daemon *Daemon) setupLinkedContainers(container *container.Container) ([]string, error) {
 	return nil, nil
-}
-
-// ConnectToNetwork connects a container to a network
-func (daemon *Daemon) ConnectToNetwork(container *container.Container, idOrName string, endpointConfig *networktypes.EndpointSettings) error {
-	return fmt.Errorf("Windows does not support connecting a running container to a network")
-}
-
-// DisconnectFromNetwork disconnects container from a network.
-func (daemon *Daemon) DisconnectFromNetwork(container *container.Container, networkName string, force bool) error {
-	return fmt.Errorf("Windows does not support disconnecting a running container from a network")
 }
 
 // getSize returns real size & virtual size
@@ -57,4 +42,8 @@ func isLinkable(child *container.Container) bool {
 
 func enableIPOnPredefinedNetwork() bool {
 	return true
+}
+
+func (daemon *Daemon) isNetworkHotPluggable() bool {
+	return false
 }
